% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_query.R
\name{db_query}
\alias{db_query}
\title{db_query}
\usage{
db_query(query, db)
}
\arguments{
\item{query}{SQL query to run}

\item{db}{the database to connect to}

\item{tbl}{logical whether to return a tibble (T) or a data frame (F)}
}
\value{
data.frame
}
\description{
creates a connection to the DB and runs SQL query
}
